+++ build.rs
@@ -23,12 +23,10 @@ fn main() {
     build.file("harfbuzz/src/hb-aat-map.cc");
     build.file("harfbuzz/src/hb-blob.cc");
     build.file("harfbuzz/src/hb-buffer.cc");
-    build.file("harfbuzz/src/hb-buffer-serialize.cc");
     build.file("harfbuzz/src/hb-common.cc");
     build.file("harfbuzz/src/hb-face.cc");
     build.file("harfbuzz/src/hb-font.cc");
     build.file("harfbuzz/src/hb-map.cc");
-    build.file("harfbuzz/src/hb-number.cc");
     build.file("harfbuzz/src/hb-ot-face.cc");
     build.file("harfbuzz/src/hb-ot-font.cc");
     build.file("harfbuzz/src/hb-ot-layout.cc");