+++ harfbuzz/src/hb-buffer.hh
@@ -56,8 +56,6 @@ static_assert((sizeof(hb_glyph_info_t) == 20), "");
 static_assert((sizeof(hb_glyph_info_t) == sizeof(hb_glyph_position_t)), "");
 
 HB_MARK_AS_FLAG_T(hb_buffer_flags_t);
-HB_MARK_AS_FLAG_T(hb_buffer_serialize_flags_t);
-HB_MARK_AS_FLAG_T(hb_buffer_diff_flags_t);
 
 enum hb_buffer_scratch_flags_t {
     HB_BUFFER_SCRATCH_FLAG_DEFAULT = 0x00000000u,