+++ src/lib.rs
@@ -20,8 +20,7 @@ pub use ttf_parser::Tag;
 
 pub use crate::buffer::{
     GlyphPosition, GlyphInfo, BufferClusterLevel,
-    SerializeFlags, SerializeFormat, BufferSerializer,
-    UnicodeBuffer, GlyphBuffer
+    SerializeFlags, UnicodeBuffer, GlyphBuffer
 };
 pub use crate::common::{Direction, Script, Language, Feature, Variation, script};
 pub use crate::font::Font;